+++ tile/kernel/memory.h
@@ -19,7 +19,7 @@ const extern uint32_t* data_end;
 const extern uint32_t* bss_begin;
 const extern uint32_t* bss_end;
 
-extern uint32_t virt_to_phys(uint32_t x);
+extern uint64_t virt_to_phys(uint32_t x);
 extern uint32_t phys_to_virt(uint32_t x);
 
 extern struct memory_map memory_map;
@@ -37,8 +37,8 @@ enum memory_map_block_flags {
   struct memory_map_block represents the bounds of a memory region.
 */
 struct memory_map_block {
-  uint32_t begin;
-  uint32_t size;
+  uint64_t begin;
+  uint64_t size;
   int flags;
 };
 
@@ -74,8 +74,8 @@ void init_memory_manager(void* pg_dir, void* text_begin, void* text_end, void* d
 
 void memory_map_mask_block(struct memory_map_block* block, int flag, int mask);
 void memory_map_merge_blocks(struct memory_map_group* group, int begin, int end);
-void memory_map_insert_block(struct memory_map_group* group, int pos, uint32_t begin, uint32_t size);
-void memory_map_add_block(struct memory_map_group* group, uint32_t begin, uint32_t size);
+void memory_map_insert_block(struct memory_map_group* group, int pos, uint64_t begin, uint64_t size);
+void memory_map_add_block(struct memory_map_group* group, uint64_t begin, uint64_t size);
 
 void* memory_alloc(size_t size);
 int memory_free(void* ptr);
