+++ tile/kernel/page.h
@@ -14,7 +14,7 @@ extern void invalidate_entire_tlb();
 uint32_t* pgd_offset(uint32_t* pgd, uint32_t addr);
 uint32_t* pte_offset(uint32_t* pte, uint32_t addr);
 uint32_t* pte_alloc(struct memory_manager* mm, uint32_t addr);
-void pte_insert(uint32_t* pte, uint32_t v_addr, uint32_t p_addr);
+void pte_insert(uint32_t* pte, uint32_t v_addr, uint64_t p_addr);
 
 void init_paging();
